+++ packages/polaris-viz-core/src/index.ts
@@ -99,6 +99,7 @@ export {
   changeColorOpacity,
   changeGradientOpacity,
   getAverageColor,
+  getValueFromXScale,
 } from './utilities';
 export {
   useSparkBar,